requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
djangorestframework==3.9.0
idna==2.8
mysql==0.0.2
mysqlclient==1.3.14
permission==0.4.1
Pygments==2.3.1
pytz==2018.9
requests==2.21.0
retrying==1.3.3
six==1.12.0
urllib3==1.24.1
